% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_wilcox.R
\name{run_de_test}
\alias{run_de_test}
\title{Run a Wilcox Test}
\usage{
run_de_test(vals, groupA, groupB, pairedFlag = FALSE)
}
\arguments{
\item{vals}{Numeric vector with sample names as keys}

\item{groupA}{Sample names from Group A}

\item{groupB}{Sample names from Group B}

\item{pairedFlag}{Run a paired wilcox mode}
}
\description{
Run a Wilcox Test
}

